Changed naming convention of command line options.
Updated nova installer to look for a saved config file named "config" in the local dir of the installer. This allows for the file to have key=value pairs for use instead of adding the values to the command line.

python nova.py --manifest_host="http://localhost/" ubuntu:10.10

is the same as the following if config is present and contains

manifest_host="http://localhost"

python nova.py ubuntu:10.10